llmops_api/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

importlib==1.0.4
injector==0.22.0
openai==1.91.0
pytest==8.4.1
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.41

# WARNING(pigar): some manual fixes might be required as pigar has detected duplicate requirements for the same import name (possibly for different submodules).
# WARNING(pigar): the following duplicate requirements are for the import name: dataclasses
dataclasses==0.8
FastFlask==1.2.32
# WARNING(pigar): the following duplicate requirements are for the import name: datetime
PyPerspective==1.0.4
sldt==1.0.4
vk-audio==6.9
# WARNING(pigar): the following duplicate requirements are for the import name: enum
enum==0.4.7
enum34==1.1.10
funutils==0.0.1
# WARNING(pigar): the following duplicate requirements are for the import name: flask
ApiBot==0.0.4
Dirtrav==1.0.0
facevision==0.0.4
FastFlask==1.2.32
Flask==3.1.1
flask-admin.py==0.1.3
Flask-Mirror-Upstream==2.3.4
flask-tmpl==0.2.0
netoprmgr==1.3.5
sahilsHousePredicition==0.0.1
sep10==0.0.3
# WARNING(pigar): the following duplicate requirements are for the import name: flask_sqlalchemy
Flask-SQLAlchemy==3.1.1
Flask-SQLAlchemy-Meiqia==2016.8.1
netoprmgr==1.3.5
# WARNING(pigar): the following duplicate requirements are for the import name: flask_wtf
# WARNING(pigar): the following duplicate requirements are for the import name: typing
typeffect==0.0.1
typing==3.10.0.0
# WARNING(pigar): the following duplicate requirements are for the import name: uuid
lib-uuid-porting==2.6
uuid==1.30
# WARNING(pigar): the following duplicate requirements are for the import name: wtforms
e-WTForms==0.0.1
netoprmgr==1.3.5
WTForms==3.2.1
